+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Both colorscad and 3mfmerge changes are included here. Unless explicitly mention
 ### Fixed
 
 - Remove dependency on GNU coreutils on macOS; now OpenSCAD is the only dependency, no more need to brew
+- Robustly support input (-i) that's not in the current dir; work around inconsistent behavior among OpenSCAD versions
 
 ## [0.5.0] - 2021-08-19